requirements.txt
django==1.8.3
jsonfield==2.0.2
MySQL-python==1.2.5
django_cleanup==1.0.1
django-suit==0.2.25
Pillow==4.2.1
django-uuslug==1.1.8
djangomako==1.1.0
djangorestframework==3.3.3
gunicorn